+++ src/main/java/epn/edu/ec/servicios/AdolescenteInfractorUDIServicio.java
@@ -68,13 +68,13 @@ public class AdolescenteInfractorUDIServicio {
         return listaAdolescentesUDIAux;
     }
 
-    public int eliminarAdolescenteInfractor(Integer id) {
-
-        int statusRespuesta = 0;
-        Response response = conexion.conexion(URL_ADOLESCENTE_INFRACTOR_UDI, "DELETE", true, null);
-        if (response.getStatus() == 200) {
-            statusRespuesta = 200;
-        }
-        return statusRespuesta;
-    }
+//    public int eliminarAdolescenteInfractor(Integer id) {
+//
+//        int statusRespuesta = 0;
+//        Response response = conexion.conexion(URL_ADOLESCENTE_INFRACTOR_UDI, "DELETE", true, null);
+//        if (response.getStatus() == 200) {
+//            statusRespuesta = 200;
+//        }
+//        return statusRespuesta;
+//    }
 }
