+++ src/main/webapp/plantillas/admin/reporte/reporteEdadFechaFuturaUDI.xhtml
@@ -11,7 +11,7 @@
             <b:form id="form">
                 <b:row>
                     <b:column span="4" offset="1">                                        
-                        <h:outputLabel for="txtFechaNacimiento" value="Fecha Nacimiento"/>
+                        <h:outputLabel for="txtFechaNacimiento" value="Fecha Futura"/>
                         <b:dateTimePicker id="txtFechaNacimiento" placeholder="Fecha" showTime="false" value="#{panelReportesControlador.fechaBuscar}"/>
                     </b:column>
                     <br/>
