+++ src/main/webapp/plantillas/admin/reporte/reporteTipoDelitoCAI.xhtml
@@ -12,7 +12,7 @@
                 <b:row>
                     <b:column span="12" offset="0">
                         <h:outputLabel for="txtTipoInfraccion" value="Tipo Infracción"/>
-                        <b:selectOneMenu id="txtTipoInfraccion" disabled="#{!detalleInfraccionControlador.guardado}" select2="true" style="width:100%" value="#{panelReportesControlador.reporte1.tipoDelto}">
+                        <b:selectOneMenu id="txtTipoInfraccion" select2="true" style="width:100%" value="#{panelReportesControlador.reporte1.tipoDelto}">
                             <f:selectItem itemLabel="Seleccione una opcion"  itemDisabled="false" />
                             <f:selectItems value="#{panelReportesControlador.tiposPenal}" var="tipos" itemLabel="#{tipos.tipoPenal}" itemValue="#{tipos.tipoPenal}"/>
                         </b:selectOneMenu>
